# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ISRO ISTRAC SOC AI Platform",
    "description": "\n        AI-Driven Security Analytics Platform for ISRO ISTRAC Bengaluru.\n\n        ## Architecture\n        This platform ingests security logs from Elasticsearch, detects anomalies\n        using ML models (Isolation Forest, Autoencoder, LSTM), scores threats,\n        and provides natural language investigation via an SLM assistant.\n\n        ## Authentication\n        All endpoints require Bearer token authentication (see /api/auth/login).\n\n        ## Rate Limits\n        - /api/slm/chat: 60 requests/minute\n        - /api/alerts: 300 requests/minute\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/training/initial": {
      "post": {
        "tags": [
          "Training"
        ],
        "summary": "Trigger Initial Training",
        "operationId": "trigger_initial_training_api_training_initial_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/incremental": {
      "post": {
        "tags": [
          "Training"
        ],
        "summary": "Trigger Incremental Retraining",
        "operationId": "trigger_incremental_retraining_api_training_incremental_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/status/{job_id}": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Job Status",
        "operationId": "get_job_status_api_training_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/status": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Training Status",
        "operationId": "get_training_status_api_training_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/history": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Training History",
        "operationId": "get_training_history_api_training_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/drift": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Drift Status",
        "operationId": "get_drift_status_api_training_drift_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/interpretability": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Interpretability Report",
        "operationId": "get_interpretability_report_api_training_interpretability_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/accuracy": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Accuracy Report",
        "operationId": "get_accuracy_report_api_training_accuracy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/threshold": {
      "post": {
        "tags": [
          "Training"
        ],
        "summary": "Update Threshold",
        "operationId": "update_threshold_api_training_threshold_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThresholdUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/mlflow/experiments": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "List Experiments",
        "operationId": "list_experiments_api_training_mlflow_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/training/mlflow/runs/{experiment_id}": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_training_mlflow_runs__experiment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/mlflow/runs/detail/{run_id}": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Run Details",
        "operationId": "get_run_details_api_training_mlflow_runs_detail__run_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/mlflow/compare": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Compare Runs",
        "operationId": "compare_runs_api_training_mlflow_compare_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Ids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/calibration": {
      "get": {
        "tags": [
          "Training"
        ],
        "summary": "Get Calibration Stats",
        "operationId": "get_calibration_stats_api_training_calibration_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Training"
        ],
        "summary": "Trigger Calibration Training",
        "operationId": "trigger_calibration_training_api_training_calibration_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ThresholdUpdate": {
        "properties": {
          "threshold": {
            "type": "number",
            "title": "Threshold"
          }
        },
        "type": "object",
        "required": [
          "threshold"
        ],
        "title": "ThresholdUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}